+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 VMware Aria Operations Integration SDK
 --------------------------------------
 ## Unreleased
+* Default container registry can now be specified in global config file
 * Fixes issue where psutil version would cause python build to fail
 * Standardize encoding of properties file to 'utf-8'
 * Adds two flags to mp-test for controlling collection parameters: